backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
neo4j==5.16.0
google-generativeai>=0.7.0
python-dotenv==1.0.0
pydantic==2.5.3
# UPDATED VERSIONS BELOW
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0